apps/PythonFastApi/requirements.txt
fastapi[standard]
psycopg[binary,pool]
uvicorn[standard]
python-dotenv
motor[srv]>=3.6
